1.0.0: a few backwards-compatible changes.
- Use standard naming conventions (which weren’t around when I first
  wrote this): rename `src/phantom_enum.rs` to `src/lib.rs`.
- Support a trailing comma in the `phantom_enum!` macro invocation.
  (All the best things support trailing commas. All the best things
  *use* trailing commas.)
- Improve the test to add bounds on the struct (this wasn’t possible in
  Rust last time I updated the code).

This is backwards-compatible with 0.9.5. I just want to make it 1.0.0.
(That was, after all, the intention at 0.9.0: make it 1.0.0 shortly. I
just took rather a long time after 0.9.5 to get round to doing it. Over
a year. Stabilising it before Rust 1.0.0 *would* have been premature.)

Woo! Another library (even if it’s rather trivial) reaches stability!

An interesting piece of future work might be making it non-phantom, and
supporting struct/tuple variants. That could be fun. But it would hardly
be a *phantom* enum then. Just a kind of type enum. So I guess it won’t
go in the phantom-enum crate.

Honestly, this crate is complete. It’s unlikely to ever have any more
work done on it because it doesn’t *need* any more work done on it.